run/helloworld/requirements.txt
gunicorn==20.1.0
Flask==2.1.1
numpy==1.23.0
Pillow==9.2.0
google-cloud-aiplatform==1.15.0
